Verify and fix remaining review findings
Re-check the reported findings against the current branch state and fix only the ones that still reproduce.

Add parser validation for declared body lengths in Hotline assembly payloads, move the message-assembly tests into a dedicated sibling module, and update the touched documentation to reflect the delivered transport behaviour and neutral British-English prose.

Leave the reported zero-length continuation change out because the current physical Hotline frame validation already rejects that input before the tracker can observe it.
