+++ src/wireframe/codec/tests.rs
@@ -277,7 +277,6 @@ fn fragment_ranges_cover_payload(#[case] total_len: usize, #[case] expected_coun
         assert!(last_len > 0);
     }
 }
-
 #[rstest]
 #[case::size_mismatch(
     FrameHeader {
